+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to `klarnaordermanagement` will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
+### Added
+- Allowed manual and individual adding of positions which are not related to an article
 
 ## [0.5.0] - 2022-02-28 
 ### Fixed