+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to `klarnaordermanagement` will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
+### Fixed
+- Fixed a missing auth header during customer token generation
+- Fixed the config reader to return the correct value when falsy values are configured (e.g. 0, 'false')
 ### Added
 - Allowed manual and individual adding of positions which are not related to an article